requirements.txt
# Core dependencies
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.2.0
requests>=2.31.0
aiohttp>=3.9.0

# LangChain and related libraries
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
langgraph>=0.0.20

# Development tools
black>=23.0.0
isort>=5.0.0

# Streamlit for UI
streamlit>=1.30.0

# Testing
pytest 
pytest-cov

